Move wlr_types/box.py to util/box.py
wlroots 0.15.0 moves their box.h so we should move ours accordingly. A
`DeprecationWarning` is emitted when importing from the current
location.
